requirements.txt
# Core dependencies
requests>=2.31.0
websockets==15.0.1
aiohttp==3.11.18
python-dotenv>=1.0.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Logging and monitoring
loguru>=0.7.0

# Database
# sqlite3은 Python 내장

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Type hints
mypy>=1.5.0

# Additional dependencies
werkzeug==3.1.3
wheel==0.45.1
win32-setctime==1.2.0
yarl==1.20.0
zipp==3.21.0
